chore : gpt2 hate speech directory renaming, latent hatred small refactor + add __init__.py
