+++ .gitignore
@@ -6,6 +6,7 @@ ENV/
 env.bak/
 venv.bak/
 
+/gpt2_hate_speech/weights
 __pycache__/
 *.py[cod]
 *$py.class
